+++ CHANGELOG.md
@@ -1,11 +1,25 @@
 ## Change log
 ----------------------
 
+Version 1.3
+-------------
+
+ADDED:
+
+- gradle as new build system
+
+CHANGED:
+
+- changed to new package io.github.astrapi69
+- update of dependency lombok to new version 1.18.18
+- update of dependency commons-lang3 version to 3.12.0
+- update of test dependency testng to new version 7.4.0
+
 Version 1.2
 -------------
 
 ADDED:
- 
+
 - this changelog file
 - new eclipse launch script for display dependency updates created
 - new license file added
@@ -16,4 +30,5 @@ CHANGED:
 - update of commons-lang3 version to 3.8.1
 
 Notable links:
-[keep a changelog](http://keepachangelog.com/en/1.0.0/) Don’t let your friends dump git logs into changelogs
+[keep a changelog](http://keepachangelog.com/en/1.0.0/) Don’t let your friends dump git logs into
+changelogs